chore(protocol): `latitude` and `longitude` exist together
While reading through this, I assumed this was a bug.
I cannot imagine a `latitude` (and `accuracyRadius`) to exist but
no `longitude`, and inverse?
Especially because these *are* checking for `accuracyRadius`,
and there is also a `hasAccuracyRadius` which *does* have a predicate
that if it exists,
both `latitude` and `longitude` exist too.

This is also what I saw in the data,
though I have not seen too much data.