fix: support passing `ip` to `protect` in more integrations
This adds support in types and in actual code for passing `ip` in `request` to
`protect`.
This applies to the astro, bun, deno, fastify, and node integrations.
The nest and next integrations already support this.
Remix and Sveltekit have custom ways to pass IPs in and do not receive a regular
`request` so I left them as-is.

This method is documented on the website at:
<https://docs.arcjet.com/troubleshooting#override-development-ip>.

Closes GH-4808.