feat(arcjet): allow calling `shield` w/o options
The shield rule accepts an options object with a single optional field `mode`,
defaulting to `DRY_RUN`.
As all fields in the options object are optional, humans who want that, now
have to pass an empty object: `shield({})`.
It is nicer for humans to be able to omit that and write `shield()`.

Closes GH-4560.