chore(tsconfig): remove `@arcjet/tsconfig`
See the related issue for more reasoning.

After this lands it remains to deprecate the package on npm.
That can be done when the next SDK is released.
At that same version, we can publish an empty package
and immediately deprecate it.
If anyone uses it, they will not automatically pull it in
with semver (because we’re at beta releases) but when they
pull it in they will see the deprecation.

Closes GH-4591.