docs(arcjet): add JSDocs (#4997)
Closes GH-4962.

This adds JSDocs to all types that are exposed.
It importantly documents all fields, parameters, etcetera on those types too.

This is useful if one for example writes `detectBot({allow: ""})`, where TypeScript now shows documentation for that allow option. To not be too verbose, functions themselves only describe their parameters, not the fields *in* those parameters.