refactor: simplify `withClient` types in adapters
This removes a gnarly type problem, which was solved with a cast,
by simplifying the types and well, sadly, still having a type cast.

But the new type cast is on a much smaller value, the default.
And the rest of the code is simpler now, so I think it’s good to have.

Related-to GH-31.