refactor(redact)!: improve types around custom entities (#4883)
The `redact` function accepts two type parameters.
The second type parameter `CustomEntities` is not used. Looking at the tests (or the examples in #4982). It is always `string`. Never something more specific.
The first type parameter, `Detect`, is an entire function, whose variadic part is that unused `CustomEntities`.
Still, it is accepted as a `const` and used in `RedactOptions`, where it does work to influence the `entities` and `replace` options.
However, it is not *needed* to put an entire function into a type parameter, if only a small part is variadic. This is not a real issue. Just a tiny bit smelly.

Still, there are some cases where type errors happen for valid use cases. This PR changes the type parameters and internal code to allow these valid use cases. With tests.

Closes GH-4982.