build(analyze-wasm, redact-wasm): use `jco-transpile`
Since there is now a `jco-transpile` package with few dependencies (for 2 weeks),
and that we’ve been stuck on an older `jco` to prevent package bloat,
this PR attempts to unblock us by switching to the new package.

It seems to work! Their package is at `0.0.4` though, and the types a bit broken, but things pass.
All options except for `no-wasi-shim` that were used are also in `jco-transpile.
I think that’s something that is wrapped around `transpile`.
It could be some shim is injected though, I do not know enough about webassembly
to verify whether some shim is in here.

Related-to: bytecodealliance/jco#500.
Related-to: GH-3446.
Related-to: GH-4342.
Closes GH-5101.