wip: proof of concept to enable `ignore-scripts`
This is a draft so that it can be discussed whether the trade offs are
worth it.

Related reading:

* <https://socket.dev/blog/pnpm-10-0-0-blocks-lifecycle-scripts-by-default>
* <https://pnpm.io/supply-chain-security>

Adding `ignore-scripts` turns of *all* pre/post (lifecycle) scripts.
In dependencies that are installed and also in the packages developed here.
Hence this now documents to run `npm run build` after `npm ci`.

Personally, I have added `ignore-scripts` to all my packages several
years ago, and always use `npm test` to make sure *everything* is perfect:
properly built, linted, and tested.