types(fastify): use minimum request interface
Updating `fastify` keeps failing the example as that is pinned
to a different one,
which are not exactly the same.

Other integrations define their own request as a minimum viable
interface.
This PR adds that too for Fastify.

I did notice that `protocol` was also on there,
and the current detection was a bit weird,
so this now uses that,
if available.

Related-to: GH-5237.