fix: support ESM types in dependencies (#5269)
While this repo uses ESM everywhere, this was not really enabled in TypeScript, as that used `moduleResolution: "node"` (an alias for `node10`), which only really supports `require` and importing files with paths.

See: <https://www.typescriptlang.org/tsconfig/#moduleResolution>.

This surfaces for example in Astro using export maps to point people to their Zod in GH-5266. That probably surfaces because of withastro/astro@ad26fa22.

As the word is `node`, we could use `nodenext` (alias `node16`), but as that would require that `module: "esnext"` be changed to something Node-specific too, using the modern `moduleResolution: "bundler"` is likely a better idea. That value does more than is needed for export maps, but it at least does not force more Node-specific configuration.

The next problem is that `@rollup/plugin-typescript` has bugs. which are not very visible because they use a stale-bot to close issues. See rollup/plugins#1583.
For that reason, these `moduleResolution` and `module` fields are explicit in `rollup-config/index.js` too.

The `bundler` module resolution algorithm slightly changes the handling of the internal types, which is why `analyze-wasm` and `redact-wasm` now reference the `wasm.d.ts` files, which importantly ***does not*** leak into the generated `.d.ts` files.

Related-to: GH-5266.