fix(ip): throw on proxy CIDR range in string form
It is also possible to throw this error in `isTrustedProxy`.
The downside of that is that the check then occurs for every potential IP,
and that the error would not happen if no potential IP is found.
This verifies the input options once.

An alternative approach is to call `parseProxy` instead of throwing.
The downside of that is that it would be faster if users do that initially
and once.

Closes GH-5382.