chore(env)!: remove unused `apiKey` utility function
This function is unused.
All SDKs require users to retrieve API keys themselves
and pass them explicitly.
The only exception is the Arcjet Astro SDK,
which does not use this function but retrieves the key
the regular Astro way.
As this function is unused and undocumented (on the site,
in examples),
and probably also not recommended for use,
I’d recommend removing it.

Closes GH-5453.