feat: send extra `vercel-git-commit-sha` field to cloud api
In `@arcjet/next`, some headers and an env variable were sent in `extra` on Vercel.
As `arcjet` does not have access to the environment, such behavior must be in integrations.
This PR changes each integration to send such information on Vercel.

As headers are already sent in headers, I removed that, but that can be added.

Related-to: GH-5456.